+++ plugins/commands/plugin/action.rb
@@ -39,12 +39,23 @@ module VagrantPlugins
         end
       end
 
+      # This middleware sequence will update a plugin.
+      def self.action_update
+        Vagrant::Action::Builder.new.tap do |b|
+          b.use BundlerCheck
+          b.use PluginExistsCheck
+          b.use InstallGem
+          b.use PruneGems
+        end
+      end
+
       # The autoload farm
       action_root = Pathname.new(File.expand_path("../action", __FILE__))
       autoload :BundlerCheck, action_root.join("bundler_check")
       autoload :InstallGem, action_root.join("install_gem")
       autoload :LicensePlugin, action_root.join("license_plugin")
       autoload :ListPlugins, action_root.join("list_plugins")
+      autoload :PluginExistsCheck, action_root.join("plugin_exists_check")
       autoload :PruneGems, action_root.join("prune_gems")
       autoload :UninstallPlugin, action_root.join("uninstall_plugin")
     end